修改上传文件bug

+++ src/Servlet/userUpdateServlet.java
@@ -32,36 +32,44 @@ public class userUpdateServlet extends HttpServlet {
         String pwdQuestion = request.getParameter("pwdQuestion");
         String pwdQuestionAnswer = request.getParameter("pwdQuestionAnswer");
         String personalSignature = request.getParameter("personalSignature");
-        Part userImg = request.getPart("userImg");
         UserService userService = new UserService();
         User user = userService.checkUser(userName);
 
         // 2.接收并保存用户头像到文件中
+        Part userImg = request.getPart("userImg");
         String userImgName = userImg.getSubmittedFileName();
-        String ext = userImgName.substring(userImgName.lastIndexOf("."));
-        // 2.1使用随机数重新赋名
-        userImgName = UUID.randomUUID() + ext;
-        String dir = getServletContext().getRealPath("userImg");
-        // 2.2判断用户是否已经有了头像,如果已经有了则删除原来的头像(节省空间)
-        if (user.getUserImgPath() != null) {
-            String oldUserImgName = user.getUserImgPath().contains("/") ?
-                    user.getUserImgPath().substring(user.getUserImgPath().lastIndexOf("/") + 1) :
-                    user.getUserImgPath().substring(user.getUserImgPath().lastIndexOf("\\") + 1);
-            // 此处使用的是java.io.File
-            File file = new File(dir + "/" + oldUserImgName);
-            // file.deleteOnExit();  这个只有重启服务器才会删除
-            Files.deleteIfExists(file.toPath()); // 这个是立即删除
+        String UUIDUserImgName = null;
+        // 判断是否上传了照片
+        if (!userImgName.isEmpty()) {
+            // 2.1获取头像后缀名
+            String ext = userImgName.substring(userImgName.lastIndexOf("."));
+            // 2.2使用随机数给头像文件重新赋名
+            UUIDUserImgName = UUID.randomUUID() + ext;
+            String dir = getServletContext().getRealPath("userImg");
+            // 2.3判断用户是否已经有了头像（是否是第一次上传）,如果已经有了则删除原来的头像(节省空间)
+            if (user.getUserImgPath() != null) {
+                String oldUserImgName = user.getUserImgPath().contains("/") ?
+                        user.getUserImgPath().substring(user.getUserImgPath().lastIndexOf("/") + 1) :
+                        user.getUserImgPath().substring(user.getUserImgPath().lastIndexOf("\\") + 1);
+                // 此处使用的是java.io.File
+                File file = new File(dir + "/" + oldUserImgName);
+                // file.deleteOnExit();  这个只有重启服务器才会删除
+                Files.deleteIfExists(file.toPath()); // 这个是立即删除
+            }
+            // Windows中文件路径是以'\'隔开，在字符串中表达则写两个，即'\\'
+            // String savePath = dir + "\\" + userImgPath;
+            // Linux中文件路径是以'/'隔开，在字符串中表达只需写一个，即'/'
+            String savePath = dir + "/" + UUIDUserImgName;
+            // 2.4保存图片
+            userImg.write(savePath);
+        } else {
+            // 如果没有重新上传，头像名称不变
+            UUIDUserImgName = user.getUserImgPath().substring(user.getUserImgPath().lastIndexOf("/") + 1);
         }
-        // Windows中文件路径是以'\'隔开，在字符串中表达则写两个，即'\\'
-        // String savePath = dir + "\\" + userImgPath;
-        // Linux中文件路径是以'/'隔开，在字符串中表达只需写一个，即'/'
-        String savePath = dir + "/" + userImgName;
-        // 2.3保存图片
-        userImg.write(savePath);
 
         // 3.传递数据到数据库中
         int userId = user.getUserId();
-        user = new User(userId, userName, userGender, userEmail, userTel, userAddress, "userImg/" + userImgName, personalSignature, pwdQuestion, pwdQuestionAnswer);
+        user = new User(userId, userName, userGender, userEmail, userTel, userAddress, "userImg/" + UUIDUserImgName, personalSignature, pwdQuestion, pwdQuestionAnswer);
         boolean b = userService.updateUser(user);
 
         // 4.每次更新user数据，重新更新session中的user值
